requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.2.2
psycopg[binary]>=3.2.1
pandas>=2.0.0
pyarrow>=14.0.0
numpy>=2.0.0
scikit-learn==1.5.2
xgboost==2.1.1
joblib>=1.4.2
pytrends>=1.5.0
feedparser>=6.0.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4